// FE/src/models/saas-feature.model.ts
export enum SaasFeature {
  POS = 'POS',
  MENU = 'MENU',
  PAYMENTS = 'PAYMENTS',
  BASIC_REPORTS = 'BASIC_REPORTS',
  STORE_SETTINGS = 'STORE_SETTINGS',
  KITCHEN = 'KITCHEN',
  WAREHOUSE = 'WAREHOUSE',
  RECIPE = 'RECIPE',
  /** Store+ */
  SHIFT_MGMT = 'SHIFT_MGMT',
  ACCOUNTING = 'ACCOUNTING',
  EMPLOYEES = 'EMPLOYEES',
  FULL_DASHBOARD = 'FULL_DASHBOARD',
  MULTI_BRANCH = 'MULTI_BRANCH',
  CRM = 'CRM',
  LOYALTY = 'LOYALTY',
  MARKETING = 'MARKETING',
  API_INTEGRATION = 'API_INTEGRATION',
  KIOSK = 'KIOSK',
  ADVANCED_ANALYTICS = 'ADVANCED_ANALYTICS',
}